deleteTransaction (User)

+++ code/controllers/controller.js
@@ -393,17 +393,33 @@ export const getTransactionsByGroupByCategory = async (req, res) => {
  */
 export const deleteTransaction = async (req, res) => {
     try {
-        const cookie = req.cookies
+        //const cookie = req.cookies;
         /*if (!cookie.accessToken) {
             return res.status(401).json({ message: "Unauthorized" }) // unauthorized
         }*/
 
-        
+        // /users/:username/transactions
+        const info = { authType: "User", username: req.url.substring(7, req.url.length - 13) };
 
-        let data = await transactions.deleteOne({ _id: req.body._id });
-        return res.json("deleted");
+        const user = await User.findOne({ username: info.username });
+        if (!user)
+            throw new Error("User not found");
+
+        const idTransaction = req.body._id;
+        const transaction = await transactions.findOne({ _id: idTransaction });
+        if (!transaction)
+            throw new Error("Transaction not found");
+
+        if (verifyAuth(req, res, info)) {
+            let data = await transactions.deleteOne({ _id: req.body._id });
+            return res.json({ message: "Your transaction has been deleted successfully" });
+        }
+
     } catch (error) {
-        res.status(400).json({ error: error.message })
+        if (error.message == "User not found" || error.message == "Transaction not found")
+            res.status(401).json({ error: error.message })
+        else
+            res.status(400).json({ error: error.message })
     }
 }
 
@@ -415,8 +431,18 @@ export const deleteTransaction = async (req, res) => {
     - error 401 is returned if at least one of the `_ids` does not have a corresponding transaction. Transactions that have an id are not deleted in this case
  */
 export const deleteTransactions = async (req, res) => {
-    try {
+    /*try {
+        // /transactions
+        const info = { authType: "Admin" };
+        let selectedTransactions = [];
+
+        if (verifyAuth(req, res, info)) {
+            await req.body._ids.forEach(element => {
+                selectedTransactions.push(transactions.findOne({ _id: element }))
+            });
+            res.status(200).json(selectedTransactions)
+        }
     } catch (error) {
         res.status(400).json({ error: error.message })
-    }
+    }*/
 }
